Add files via upload
Problem found when files extend down into bottom half of rfs (located in the sideways ROM area) and are overwritten by a different ROM. This caused wrfs to crash when it is restored. The update2 and defrag routines have been modified to check whether a different ROM has been loaded and if so ignore any files that extend down into the sideways ROM area.